LLM/requirements.txt
anthropic>=0.32.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pillow>=10.0.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.0
requests>=2.31.0
